+++ mpaas-20201028/alibabacloud_mpaas20201028/models/_create_template_response_body.py
@@ -16,7 +16,6 @@ class CreateTemplateResponseBody(DaraModel):
         self.code = code
         self.data = data
         self.msg = msg
-        # Id of the request
         self.request_id = request_id
         self.success = success