+++ mpaas-20201028/alibabacloud_mpaas20201028/models/_get_auth_token_to_msence_response_body.py
@@ -15,7 +15,6 @@ class GetAuthTokenToMsenceResponseBody(DaraModel):
         success: bool = None,
     ):
         self.mpaas_system_oauth_token_response = mpaas_system_oauth_token_response
-        # Id of the request
         self.request_id = request_id
         self.result_code = result_code
         self.result_msg = result_msg