+++ mpaas-20201028/alibabacloud_mpaas20201028/models/_get_template_response_body.py
@@ -17,7 +17,6 @@ class GetTemplateResponseBody(DaraModel):
         self.code = code
         self.data = data
         self.msg = msg
-        # Id of the request
         self.request_id = request_id
         self.success = success