+++ mpaas-20201028/alibabacloud_mpaas20201028/models/_query_link_response_body.py
@@ -15,7 +15,6 @@ class QueryLinkResponseBody(DaraModel):
         result_content: main_models.QueryLinkResponseBodyResultContent = None,
         result_message: str = None,
     ):
-        # Id of the request
         self.request_id = request_id
         self.result_code = result_code
         self.result_content = result_content