+++ mpaas-20201028/alibabacloud_mpaas20201028/models/_update_link_response_body.py
@@ -13,7 +13,6 @@ class UpdateLinkResponseBody(DaraModel):
         result_content: main_models.UpdateLinkResponseBodyResultContent = None,
         result_message: str = None,
     ):
-        # Id of the request
         self.request_id = request_id
         self.result_code = result_code
         self.result_content = result_content