+++ mpaas-20201028/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-06-23 Version: 3.3.3
+- Update API SaveOrderRelationInfoToMsence: add request parameters ItemId.
+- Update API SaveOrderRelationInfoToMsence: add request parameters ItemTitle.
+
+
 2026-06-08 Version: 3.3.2
 - Update API SaveOrderRelationInfoToMsence: add request parameters ClientType.
 - Update API SaveOrderRelationInfoToMsence: add request parameters CpExtra.